project/requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
httpx
pytest
psycopg[binary]
redis
pika
